fix: edge case robustness, docs, and C++ overflow bugfixes
- Stats.R: fix Froh_inbreeding crash on empty runs (genome_wide T and F)
  and missing-individuals bug (use sample_info not runs rows for base table);
  fix tableRuns progress-bar crash (max=0) and seq(1,0) off-by-one;
  fix summaryRuns crash on empty runs via early return after Froh
- binary_roh.R: allow saveROH on PED-engine ROH objects (synthesize chrom_map)
- output.cpp: fix uint8_t overflow in write_roh_binary chrom name lookup
  (static_cast<uint8_t>(256)==0 broke every chrom name written to binary)
- rcpp_entry.cpp: explicit size_t casts for chrom_names array indexing
- NEWS.md: document runsAssociation() under detectRUNS 1.0.0
- vignettes: add association testing section with worked sheep example
- tests/testthat/test_edge_cases.R: 72 new edge-case and stress tests
  (all pass: 163 total, 0 fail)

+++ detectRUNS/src/output.cpp
@@ -99,7 +99,7 @@ void write_roh_binary(
         // Write chrom table
         for (uint8_t cidx : chrom_idxs) {
             const std::string& name =
-                (cidx < static_cast<uint8_t>(bim.chrom_names.size()) &&
+                (static_cast<size_t>(cidx) < bim.chrom_names.size() &&
                  !bim.chrom_names[cidx].empty())
                 ? bim.chrom_names[cidx]
                 : std::to_string(cidx);
